refactor: move logout to userMenu
