+++ src/app/components/header/UserMenu.jsx
@@ -1,8 +1,11 @@
 import React, { useEffect, useRef } from "react";
 import { translate } from "react-i18next";
 import { Link } from "react-router-dom";
+import { noAuthNeeded } from "../../helpers/authenticate";
+import { getLogin } from "../../helpers/authenticate";
 import { buildClassName as cn } from "../../helpers/buildClassName";
 import { outsideClickEffect } from "../../helpers/useOutsideClick";
+import { getUserName } from "../../helpers/userNameHelper";
 import UserIcon from "../profile/UserIcon";
 
 function UserMenu({ langtag, t }) {
@@ -10,6 +13,11 @@ function UserMenu({ langtag, t }) {
   const [popupOpen, setPopup] = React.useState(false);
   const closePopup = React.useCallback(() => setPopup(false));
   const togglePopup = React.useCallback(() => setPopup(!popupOpen));
+
+  const userName = getUserName();
+  const handleLogout = React.useCallback(() => {
+    getLogin().logout();
+  });
 
   useEffect(
     outsideClickEffect({
@@ -33,6 +41,16 @@ function UserMenu({ langtag, t }) {
             {t("header:menu.profile")}
           </Link>
         </li>
+
+        {!noAuthNeeded() && (
+          <li className="user-menu__entry">
+            <button className="user-menu__button" onClick={handleLogout}>
+              <i className="fa fa-power-off" />
+              {t("header:menu.logout")}
+              <span className="user-menu__user-name">{`(${userName})`}</span>
+            </button>
+          </li>
+        )}
       </ul>
     </nav>
   );
