+++ src/scss/header.scss
@@ -284,10 +284,4 @@
       }
     }
   }
-
-  .main-navigation-entry__username {
-    color: $color-text-medium-grey;
-    font-style: italic;
-    margin-left: 5px;
-  }
 }
